abstract out chain config
